+++ packages/formik/src/connect.tsx
@@ -1,9 +1,8 @@
-import * as React from 'react';
 import hoistNonReactStatics from 'hoist-non-react-statics';
-
-import { FormikContextType } from './types';
-import { FormikConsumer } from './FormikContext';
+import * as React from 'react';
 import invariant from 'tiny-warning';
+import { FormikConsumer } from './FormikContext';
+import { FormikContextType } from './types';
 
 /**
  * Connect any component to Formik context, and inject as a prop called `formik`;
@@ -13,15 +12,15 @@ export function connect<OuterProps, Values = {}>(
   Comp: React.ComponentType<OuterProps & { formik: FormikContextType<Values> }>
 ) {
   const C: React.FC<OuterProps> = (props: OuterProps) => (
-    <FormikConsumer>
-      {formik => {
+    <FormikConsumer
+      children={(formik: FormikContextType<Values>) => {
         invariant(
           !!formik,
           `Formik context is undefined, please verify you are rendering <Form>, <Field>, <FastField>, <FieldArray>, or your custom context-using component as a child of a <Formik> component. Component name: ${Comp.name}`
         );
         return <Comp {...props} formik={formik} />;
       }}
-    </FormikConsumer>
+    />
   );
   const componentDisplayName =
     Comp.displayName ||