+++ packages/formik/test/Field.test.tsx
@@ -12,6 +12,7 @@ import {
   Field,
   FastField,
   FieldProps,
+  FastFieldProps,
   FieldConfig,
   FormikProps,
   FormikConfig,
@@ -81,7 +82,7 @@ const createRenderField = (
 };
 
 const renderField = createRenderField(Field);
-const renderFastField = createRenderField(FastField);
+const renderFastField = createRenderField(FastField as $FixMe);
 
 function cases(
   title: string,
@@ -125,16 +126,16 @@ describe('Field / FastField', () => {
         <>
           <Field name="name" children={Component} />
           <Field name="name" render={Component} />
-          <Field name="name" component={Component} />
+          <Field name="name" component={Component as $FixMe} />
           <Field name="name" as={AsComponent} />
         </>
       );
 
-      const { handleBlur, handleChange } = getFormProps();
+      const { handleBlur } = getFormProps();
       injected.forEach((props, idx) => {
         expect(props.field.name).toBe('name');
         expect(props.field.value).toBe('jared');
-        expect(props.field.onChange).toBe(handleChange);
+        expect(props.field.onChange).toEqual(expect.any(Function));
         expect(props.field.onBlur).toBe(handleBlur);
         expect(props.form).toEqual(getFormProps());
         if (idx !== 2) {
@@ -151,15 +152,15 @@ describe('Field / FastField', () => {
 
       expect(asInjectedProps.name).toBe('name');
       expect(asInjectedProps.value).toBe('jared');
-      expect(asInjectedProps.onChange).toBe(handleChange);
+      expect(asInjectedProps.onChange).toEqual(expect.any(Function));
       expect(asInjectedProps.onBlur).toBe(handleBlur);
 
       expect(queryAllByText(TEXT)).toHaveLength(4);
     });
 
     it('<FastField />', () => {
-      let injected: FieldProps[] = [];
-      let asInjectedProps: FieldProps['field'] = {} as any;
+      let injected: FastFieldProps[] = [];
+      let asInjectedProps: FastFieldProps['field'] = {} as any;
 
       const Component = (props: FieldProps) =>
         injected.push(props) && <div>{TEXT}</div>;
@@ -169,20 +170,19 @@ describe('Field / FastField', () => {
       const { getFormProps, queryAllByText } = renderForm(
         <>
           <FastField name="name" children={Component} />
-          <FastField name="name" render={Component} />
           {/* @todo fix the types here?? #shipit */}
+          <FastField name="name" render={Component as $FixMe} />
           <FastField name="name" component={Component as $FixMe} />
           <FastField name="name" as={AsComponent} />
         </>
       );
 
-      const { handleBlur, handleChange } = getFormProps();
+      const { handleBlur } = getFormProps();
       injected.forEach((props, idx) => {
         expect(props.field.name).toBe('name');
         expect(props.field.value).toBe('jared');
-        expect(props.field.onChange).toBe(handleChange);
+        expect(props.field.onChange).toEqual(expect.any(Function));
         expect(props.field.onBlur).toBe(handleBlur);
-        expect(props.form).toEqual(getFormProps());
         if (idx !== 2) {
           expect(props.meta.value).toBe('jared');
           expect(props.meta.error).toBeUndefined();
@@ -197,7 +197,7 @@ describe('Field / FastField', () => {
 
       expect(asInjectedProps.name).toBe('name');
       expect(asInjectedProps.value).toBe('jared');
-      expect(asInjectedProps.onChange).toBe(handleChange);
+      expect(asInjectedProps.onChange).toEqual(expect.any(Function));
       expect(asInjectedProps.onBlur).toBe(handleBlur);
       expect(queryAllByText(TEXT)).toHaveLength(4);
     });